A simple turn signal kit
This sketch I used to create a turn signal kit for my bicycle. This kit keeps the lights on, and blinks only when a switch is turned. I have ledP ( pin #8 ) outputting constantly into a SPDT with center off. Once I turn the switch one direction, the signal then is read into pushButtonL ( or R ) and blinks as long as the switch is letting the signal come back to the arduino from the pin. Simple yet effective.